requirements.txt
docker==5.0.3
requests==2.27.1
atlassian-python-api==3.20.1
prettytable=3.2.0
